6.2 complete
Не уверен, этого ли ожидали, но модель работает.
